only use Data.{Int,}Map.Lazy
Unfortunately it's not possible to give a separate instance for
Data.Map.Lazy and Data.Map.Strict, because there is only one Map type,
not two; it is the functions exposed by those two modules which are
lazy and strict, not the datatype itself.

So if we want to give an instance, we have to pick one of the two, Lazy
or Strict. The other instances of the Map datatype have picked Lazy, so
I follow suit.

Same for IntMap.